+++ packages/terra-arrange/src/terra-dev-site/doc/arrange/Arrange.1.doc.jsx
@@ -3,11 +3,9 @@ import DocTemplate from 'terra-doc-template';
 import ReadMe from '../../../../docs/README.md';
 import { name, version } from '../../../../package.json';
 
-/* eslint-disable import/no-webpack-loader-syntax, import/no-duplicates */
-// Component Source
-import ArrangeSrc from '!raw-loader!../../../../src/Arrange';
-
 // Example Files
+import ArrangePropsTable from '!terra-props-table-loader!../../../../src/Arrange';
+import ArrangeSrc from '!raw-loader!../../../../src/Arrange';
 import ArrangeAlignAllContainers from '../example/ArrangeAlignAllContainers';
 import ArrangeAlignAllContainersSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/ArrangeAlignAllContainers';
 import ArrangeAlignFill from '../example/ArrangeAlignFill';
@@ -16,7 +14,6 @@ import ArrangeAlignFitEnd from '../example/ArrangeAlignFitEnd';
 import ArrangeAlignFitEndSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/ArrangeAlignFitEnd';
 import ArrangeAlignFitStart from '../example/ArrangeAlignFitStart';
 import ArrangeAlignFitStartSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/ArrangeAlignFitStart';
-/* eslint-enable import/no-webpack-loader-syntax, import/no-duplicates */
 
 const DocPage = () => (
   <DocTemplate
@@ -50,6 +47,7 @@ const DocPage = () => (
       {
         componentName: 'Arrange',
         componentSrc: ArrangeSrc,
+        componentProps: ArrangePropsTable,
       },
     ]}
   />