+++ packages/terra-form-select/src/terra-dev-site/doc/form-select/MultiSelectField.08.doc.jsx
@@ -2,12 +2,12 @@
 import React from 'react';
 import DocTemplate from 'terra-doc-template';
 import ReadMe from '../../../../docs/MultiSelectField.md';
+
 import { name, version } from '../../../../package.json';
 
-// Component Source
-import MultiSelectFieldSrc from '!raw-loader!../../../../src/MultiSelectField';
-
 // Example Files
+import MultiSelectFieldPropsTable from '!terra-props-table-loader!../../../../src/MultiSelectField';
+import MultiSelectFieldSrc from '!raw-loader!../../../../src/MultiSelectField';
 import SelectFieldExample from '../example/multiple/MultiSelectField';
 import SelectFieldExampleSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/multiple/MultiSelectField';
 import RequiredMultiSelectFieldExample from '../example/multiple/RequiredMultiSelectField';
@@ -19,10 +19,13 @@ const DocPage = () => (
     packageVersion={version}
     readme={ReadMe}
     srcPath={`https://github.com/cerner/terra-core/tree/master/packages/${name}`}
-    propsTables={[{
-      componentName: 'MultiSelect Field',
-      componentSrc: MultiSelectFieldSrc,
-    }]}
+    propsTables={[
+      {
+        componentName: 'Multi Select Field',
+        componentSrc: MultiSelectFieldSrc,
+        componentProps: MultiSelectFieldPropsTable,
+      },
+    ]}
     examples={[{
       title: 'MultiSelect Field Example',
       example: <SelectFieldExample />,