+++ packages/terra-toggle-section-header/src/terra-dev-site/doc/toggle-section-header/ToggleSectionHeader.1.doc.jsx
@@ -3,10 +3,9 @@ import DocTemplate from 'terra-doc-template';
 import ReadMe from '../../../../docs/README.md';
 import { name, version } from '../../../../package.json';
 
-// Component Source
-import ToggleSectionHeaderSrc from '!raw-loader!../../../../src/ToggleSectionHeader';
-
 // Example Files
+import ToggleSectionHeaderPropsTable from '!terra-props-table-loader!../../../../src/ToggleSectionHeader';
+import ToggleSectionHeaderSrc from '!raw-loader!../../../../src/ToggleSectionHeader';
 import DefaultToggleSectionHeader from '../example/DefaultToggleSectionHeader';
 import DefaultToggleSectionHeaderSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/DefaultToggleSectionHeader';
 import AnimatedToggleSectionHeader from '../example/AnimatedToggleSectionHeader';
@@ -41,6 +40,7 @@ const DocPage = () => (
       {
         componentName: 'Toggle Section Header',
         componentSrc: ToggleSectionHeaderSrc,
+        componentProps: ToggleSectionHeaderPropsTable,
       },
     ]}
   />